add frontend, servicedb message passing done

+++ servicedb/app/models.py
@@ -1,7 +1,6 @@
-from sqlalchemy import Column, DateTime, Integer, String, Text
+from sqlalchemy import Column, DateTime, Integer, String, Text, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.sql import func
-
 
 Base = declarative_base()
 
@@ -12,6 +11,6 @@ class Appeal(Base):
     surname = Column(String)
     name = Column(String)
     patronymic = Column(String)
-    phone_number = Column(Integer)
+    phone_number = Column(BigInteger)
     description = Column(Text)
     time_created = Column(DateTime(timezone=True), server_default=func.now())
